move ConceptualWallet => wallet category
