+++ app/api/ada/lib/storage/database/wallet/api/write.js
@@ -4,11 +4,16 @@ import type {
   lf$Database,
   lf$Transaction,
 } from 'lovefield';
+import {
+  op,
+} from 'lovefield';
 
 import * as Tables from '../tables';
 import type {
+  ConceptualWalletInsert, ConceptualWalletRow,
   LastSyncInfoInsert, LastSyncInfoRow,
   HwWalletMetaInsert, HwWalletMetaRow,
+
 } from '../tables';
 
 import {
@@ -66,5 +71,50 @@ export class ModifyHwWalletMeta {
       insert,
       ModifyHwWalletMeta.ownTables[Tables.HwWalletMetaSchema.name].name,
     );
+  }
+}
+
+export class ModifyConceptualWallet {
+  static ownTables = Object.freeze({
+    [Tables.ConceptualWalletSchema.name]: Tables.ConceptualWalletSchema,
+  });
+  static depTables = Object.freeze({});
+
+  static async add(
+    db: lf$Database,
+    tx: lf$Transaction,
+    request: ConceptualWalletInsert,
+  ): Promise<ConceptualWalletRow> {
+    return await addNewRowToTable<ConceptualWalletInsert, ConceptualWalletRow>(
+      db, tx,
+      request,
+      ModifyConceptualWallet.ownTables[Tables.ConceptualWalletSchema.name].name,
+    );
+  }
+
+  static async rename(
+    db: lf$Database,
+    tx: lf$Transaction,
+    request: {
+      walletId: number,
+      newName: string,
+    },
+  ): Promise<void> {
+    const conceptualWalletTable = db.getSchema().table(
+      ModifyConceptualWallet.ownTables[Tables.ConceptualWalletSchema.name].name
+    );
+    const updateQuery = db
+      .update(conceptualWalletTable)
+      .set(
+        conceptualWalletTable[Tables.ConceptualWalletSchema.properties.Name],
+        request.newName
+      )
+      .where(op.and(
+        conceptualWalletTable[Tables.ConceptualWalletSchema.properties.ConceptualWalletId].eq(
+          request.walletId
+        ),
+      ));
+
+    await tx.attach(updateQuery);
   }
 }
